Polish file-level consistency
