+++ messages/ar.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "اضغط مطوّلًا لمسح الرسالة.",
   "messagePlay": "تشغيل الرسالة",
   "messageStop": "إيقاف",
-  "onboardingSmartRewritingTitle": "إعادة صياغة ذكية",
   "onboardingSmartRewritingDescription": "وسّع النقر على الرموز إلى جمل طبيعية واضبط النبرة: مباشرة أو احترافية أو ودّية.",
-  "onboardingTranslationTitle": "ترجمة",
   "onboardingTranslationDescription": "ترجم اللوحات محليًا للتواصل بعدة لغات.",
-  "onboardingPrivacyTitle": "خصوصية ودون اتصال",
   "onboardingPrivacyDescription": "يعمل بالكامل على جهازك دون أي تتبّع للبيانات أو اعتماد على السحابة.",
   "onboardingContinue": "متابعة",
   "errorGenericTitle": "حدث خطأ ما",
