+++ messages/bg.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Задръжте, за да изчистите съобщението.",
   "messagePlay": "Възпроизвеждане на съобщението",
   "messageStop": "Спиране",
-  "onboardingSmartRewritingTitle": "Интелигентно преписване",
   "onboardingSmartRewritingDescription": "Превърнете докосванията на плочките в естествени изречения и регулирайте тона: директен, професионален или приятелски.",
-  "onboardingTranslationTitle": "Превод",
   "onboardingTranslationDescription": "Превеждайте дъските локално, за да общувате на няколко езика.",
-  "onboardingPrivacyTitle": "Поверителност и офлайн",
   "onboardingPrivacyDescription": "Работи изцяло на вашето устройство, без проследяване на данни и без зависимост от облака.",
   "onboardingContinue": "Продължаване",
   "errorGenericTitle": "Нещо се обърка",
