+++ messages/bn.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "বার্তা মুছতে দীর্ঘক্ষণ চাপুন।",
   "messagePlay": "বার্তা চালান",
   "messageStop": "থামান",
-  "onboardingSmartRewritingTitle": "স্মার্ট পুনর্লিখন",
   "onboardingSmartRewritingDescription": "টাইলে ট্যাপকে স্বাভাবিক বাক্যে রূপান্তর করুন এবং টোন সামঞ্জস্য করুন: সরাসরি, পেশাদার বা বন্ধুত্বপূর্ণ।",
-  "onboardingTranslationTitle": "অনুবাদ",
   "onboardingTranslationDescription": "একাধিক ভাষায় যোগাযোগ করতে বোর্ড স্থানীয়ভাবে অনুবাদ করুন।",
-  "onboardingPrivacyTitle": "ব্যক্তিগত ও অফলাইন",
   "onboardingPrivacyDescription": "সম্পূর্ণরূপে আপনার ডিভাইসে চলে, কোনো ডেটা ট্র্যাকিং ছাড়াই এবং ক্লাউডের ওপর নির্ভরতা ছাড়াই।",
   "onboardingContinue": "চালিয়ে যান",
   "errorGenericTitle": "কিছু একটা ভুল হয়েছে",
