+++ messages/cs.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Dlouhým stisknutím zprávu vymažete.",
   "messagePlay": "Přehrát zprávu",
   "messageStop": "Zastavit",
-  "onboardingSmartRewritingTitle": "Chytrý přepis",
   "onboardingSmartRewritingDescription": "Rozšiřte klepnutí na dlaždice do přirozených vět a upravte tón: přímý, profesionální nebo přátelský.",
-  "onboardingTranslationTitle": "Překlad",
   "onboardingTranslationDescription": "Překládejte tabulky lokálně, abyste komunikovali ve více jazycích.",
-  "onboardingPrivacyTitle": "Soukromé a offline",
   "onboardingPrivacyDescription": "Běží zcela ve vašem zařízení, bez sledování dat a bez závislosti na cloudu.",
   "onboardingContinue": "Pokračovat",
   "errorGenericTitle": "Něco se pokazilo",
