+++ messages/da.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Hold nede for at rydde beskeden.",
   "messagePlay": "Afspil besked",
   "messageStop": "Stop",
-  "onboardingSmartRewritingTitle": "Smart omskrivning",
   "onboardingSmartRewritingDescription": "Udvid tryk på felter til naturlige sætninger, og justér tonen: direkte, professionel eller venlig.",
-  "onboardingTranslationTitle": "Oversættelse",
   "onboardingTranslationDescription": "Oversæt tavler lokalt, så du kan kommunikere på flere sprog.",
-  "onboardingPrivacyTitle": "Privat og offline",
   "onboardingPrivacyDescription": "Kører helt på din enhed, uden datasporing og uden afhængighed af skyen.",
   "onboardingContinue": "Fortsæt",
   "errorGenericTitle": "Noget gik galt",
