+++ messages/de.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Lange drücken, um die Nachricht zu löschen.",
   "messagePlay": "Nachricht abspielen",
   "messageStop": "Stopp",
-  "onboardingSmartRewritingTitle": "Intelligentes Umformulieren",
   "onboardingSmartRewritingDescription": "Erweitere Kachel-Tipps zu natürlichen Sätzen und passe den Ton an: direkt, professionell oder freundlich.",
-  "onboardingTranslationTitle": "Übersetzung",
   "onboardingTranslationDescription": "Übersetze Tafeln lokal, um in mehreren Sprachen zu kommunizieren.",
-  "onboardingPrivacyTitle": "Privat & offline",
   "onboardingPrivacyDescription": "Läuft vollständig auf deinem Gerät – ohne Datenverfolgung und ohne Cloud-Abhängigkeiten.",
   "onboardingContinue": "Weiter",
   "errorGenericTitle": "Etwas ist schiefgelaufen",
