+++ messages/el.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Πατήστε παρατεταμένα για εκκαθάριση του μηνύματος.",
   "messagePlay": "Αναπαραγωγή μηνύματος",
   "messageStop": "Διακοπή",
-  "onboardingSmartRewritingTitle": "Έξυπνη αναδιατύπωση",
   "onboardingSmartRewritingDescription": "Επεκτείνετε τα πατήματα στα πλακίδια σε φυσικές προτάσεις και προσαρμόστε τον τόνο: άμεσος, επαγγελματικός ή φιλικός.",
-  "onboardingTranslationTitle": "Μετάφραση",
   "onboardingTranslationDescription": "Μεταφράστε τους πίνακες τοπικά, για να επικοινωνείτε σε πολλές γλώσσες.",
-  "onboardingPrivacyTitle": "Ιδιωτικό και εκτός σύνδεσης",
   "onboardingPrivacyDescription": "Εκτελείται εξ ολοκλήρου στη συσκευή σας, χωρίς παρακολούθηση δεδομένων και χωρίς εξάρτηση από το cloud.",
   "onboardingContinue": "Συνέχεια",
   "errorGenericTitle": "Κάτι πήγε στραβά",
