+++ messages/en.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Long press to clear message.",
   "messagePlay": "Play message",
   "messageStop": "Stop message",
-  "onboardingSmartRewritingTitle": "Smart Rewriting",
   "onboardingSmartRewritingDescription": "Say more with fewer taps. AI helps your messages sound clear and natural.",
-  "onboardingTranslationTitle": "Translation",
   "onboardingTranslationDescription": "Communicate across languages, right from your board.",
-  "onboardingPrivacyTitle": "Private & Offline",
   "onboardingPrivacyDescription": "Your communication stays private and works offline. No subscription needed.",
   "onboardingContinue": "Continue",
   "errorGenericTitle": "Something went wrong",
