+++ messages/es.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Mantén pulsado para borrar el mensaje.",
   "messagePlay": "Reproducir mensaje",
   "messageStop": "Detener",
-  "onboardingSmartRewritingTitle": "Reescritura inteligente",
   "onboardingSmartRewritingDescription": "Convierte los toques en las casillas en frases naturales y ajusta el tono: directo, profesional o amistoso.",
-  "onboardingTranslationTitle": "Traducción",
   "onboardingTranslationDescription": "Traduce los tableros localmente para comunicarte en varios idiomas.",
-  "onboardingPrivacyTitle": "Privado y sin conexión",
   "onboardingPrivacyDescription": "Se ejecuta por completo en tu dispositivo, sin rastreo de datos ni dependencia de la nube.",
   "onboardingContinue": "Continuar",
   "errorGenericTitle": "Algo salió mal",
