+++ messages/fi.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Tyhjennä viesti painamalla pitkään.",
   "messagePlay": "Toista viesti",
   "messageStop": "Pysäytä",
-  "onboardingSmartRewritingTitle": "Älykäs uudelleenkirjoitus",
   "onboardingSmartRewritingDescription": "Laajenna ruutujen napautukset luonnollisiksi lauseiksi ja säädä sävyä: suora, ammattimainen tai ystävällinen.",
-  "onboardingTranslationTitle": "Käännös",
   "onboardingTranslationDescription": "Käännä taulut paikallisesti, jotta voit viestiä useilla kielillä.",
-  "onboardingPrivacyTitle": "Yksityinen ja offline",
   "onboardingPrivacyDescription": "Toimii kokonaan laitteellasi ilman tietojen seurantaa ja ilman pilviriippuvuutta.",
   "onboardingContinue": "Jatka",
   "errorGenericTitle": "Jokin meni pieleen",
