+++ messages/fr.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Appui long pour effacer le message.",
   "messagePlay": "Lire le message",
   "messageStop": "Arrêter",
-  "onboardingSmartRewritingTitle": "Réécriture intelligente",
   "onboardingSmartRewritingDescription": "Transformez les appuis sur les cases en phrases naturelles et ajustez le ton : direct, professionnel ou amical.",
-  "onboardingTranslationTitle": "Traduction",
   "onboardingTranslationDescription": "Traduisez les tableaux localement pour communiquer en plusieurs langues.",
-  "onboardingPrivacyTitle": "Privé et hors ligne",
   "onboardingPrivacyDescription": "Fonctionne entièrement sur votre appareil, sans suivi des données ni dépendance au cloud.",
   "onboardingContinue": "Continuer",
   "errorGenericTitle": "Une erreur s'est produite",
