+++ messages/he.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "לחיצה ארוכה לניקוי ההודעה.",
   "messagePlay": "השמעת הודעה",
   "messageStop": "עצירה",
-  "onboardingSmartRewritingTitle": "שכתוב חכם",
   "onboardingSmartRewritingDescription": "הרחיבו הקשות על משבצות למשפטים טבעיים והתאימו את הסגנון: ישיר, מקצועי או ידידותי.",
-  "onboardingTranslationTitle": "תרגום",
   "onboardingTranslationDescription": "תרגמו את הלוחות באופן מקומי כדי לתקשר במספר שפות.",
-  "onboardingPrivacyTitle": "פרטי ולא מקוון",
   "onboardingPrivacyDescription": "פועל לחלוטין על המכשיר שלך, ללא מעקב אחר נתונים וללא תלות בענן.",
   "onboardingContinue": "המשך",
   "errorGenericTitle": "משהו השתבש",
