+++ messages/hi.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "संदेश साफ़ करने के लिए देर तक दबाएँ।",
   "messagePlay": "संदेश चलाएँ",
   "messageStop": "रोकें",
-  "onboardingSmartRewritingTitle": "स्मार्ट पुनर्लेखन",
   "onboardingSmartRewritingDescription": "टाइल टैप को स्वाभाविक वाक्यों में बदलें और टोन समायोजित करें: प्रत्यक्ष, पेशेवर या मैत्रीपूर्ण।",
-  "onboardingTranslationTitle": "अनुवाद",
   "onboardingTranslationDescription": "कई भाषाओं में संवाद करने के लिए बोर्ड का स्थानीय रूप से अनुवाद करें।",
-  "onboardingPrivacyTitle": "निजी और ऑफ़लाइन",
   "onboardingPrivacyDescription": "पूरी तरह आपके डिवाइस पर चलता है, बिना किसी डेटा ट्रैकिंग और बिना क्लाउड पर निर्भरता के।",
   "onboardingContinue": "जारी रखें",
   "errorGenericTitle": "कुछ गलत हो गया",
