+++ messages/hr.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Dugo pritisnite za brisanje poruke.",
   "messagePlay": "Reproduciraj poruku",
   "messageStop": "Zaustavi",
-  "onboardingSmartRewritingTitle": "Pametno preoblikovanje",
   "onboardingSmartRewritingDescription": "Proširite dodire na pločicama u prirodne rečenice i prilagodite ton: izravni, profesionalni ili prijateljski.",
-  "onboardingTranslationTitle": "Prijevod",
   "onboardingTranslationDescription": "Prevodite ploče lokalno kako biste komunicirali na više jezika.",
-  "onboardingPrivacyTitle": "Privatno i izvanmrežno",
   "onboardingPrivacyDescription": "Radi u potpunosti na vašem uređaju, bez praćenja podataka i bez ovisnosti o oblaku.",
   "onboardingContinue": "Nastavi",
   "errorGenericTitle": "Nešto je pošlo po zlu",
