+++ messages/hu.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Tartsd lenyomva az üzenet törléséhez.",
   "messagePlay": "Üzenet lejátszása",
   "messageStop": "Leállítás",
-  "onboardingSmartRewritingTitle": "Okos átfogalmazás",
   "onboardingSmartRewritingDescription": "Bővítsd a csempékre koppintást természetes mondatokká, és állítsd be a hangnemet: közvetlen, professzionális vagy barátságos.",
-  "onboardingTranslationTitle": "Fordítás",
   "onboardingTranslationDescription": "Fordítsd le a táblákat helyben, hogy több nyelven kommunikálhass.",
-  "onboardingPrivacyTitle": "Privát és offline",
   "onboardingPrivacyDescription": "Teljes egészében az eszközödön fut, adatkövetés nélkül és felhőfüggőség nélkül.",
   "onboardingContinue": "Folytatás",
   "errorGenericTitle": "Hiba történt",
