+++ messages/id.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Tekan lama untuk menghapus pesan.",
   "messagePlay": "Putar pesan",
   "messageStop": "Hentikan",
-  "onboardingSmartRewritingTitle": "Penulisan ulang cerdas",
   "onboardingSmartRewritingDescription": "Perluas ketukan pada petak menjadi kalimat alami dan sesuaikan nada: langsung, profesional, atau ramah.",
-  "onboardingTranslationTitle": "Terjemahan",
   "onboardingTranslationDescription": "Terjemahkan papan secara lokal untuk berkomunikasi dalam berbagai bahasa.",
-  "onboardingPrivacyTitle": "Privat & offline",
   "onboardingPrivacyDescription": "Berjalan sepenuhnya di perangkat Anda, tanpa pelacakan data dan tanpa ketergantungan cloud.",
   "onboardingContinue": "Lanjutkan",
   "errorGenericTitle": "Terjadi kesalahan",
