+++ messages/it.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Tieni premuto per cancellare il messaggio.",
   "messagePlay": "Riproduci messaggio",
   "messageStop": "Interrompi",
-  "onboardingSmartRewritingTitle": "Riscrittura intelligente",
   "onboardingSmartRewritingDescription": "Trasforma i tocchi sulle caselle in frasi naturali e regola il tono: diretto, professionale o amichevole.",
-  "onboardingTranslationTitle": "Traduzione",
   "onboardingTranslationDescription": "Traduci le tabelle localmente per comunicare in più lingue.",
-  "onboardingPrivacyTitle": "Privato e offline",
   "onboardingPrivacyDescription": "Funziona interamente sul tuo dispositivo, senza tracciamento dei dati né dipendenza dal cloud.",
   "onboardingContinue": "Continua",
   "errorGenericTitle": "Qualcosa è andato storto",
