+++ messages/ja.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "長押しでメッセージを消去します。",
   "messagePlay": "メッセージを再生",
   "messageStop": "停止",
-  "onboardingSmartRewritingTitle": "スマートリライト",
   "onboardingSmartRewritingDescription": "タイルのタップを自然な文章に展開し、トーンを調整できます：率直・プロフェッショナル・フレンドリー。",
-  "onboardingTranslationTitle": "翻訳",
   "onboardingTranslationDescription": "ボードをデバイス上で翻訳し、複数の言語でコミュニケーションできます。",
-  "onboardingPrivacyTitle": "プライベート＆オフライン",
   "onboardingPrivacyDescription": "すべてお使いのデバイス上で動作し、データの追跡やクラウドへの依存は一切ありません。",
   "onboardingContinue": "続ける",
   "errorGenericTitle": "問題が発生しました",
