+++ messages/kn.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "ಸಂದೇಶವನ್ನು ತೆರವುಗೊಳಿಸಲು ದೀರ್ಘವಾಗಿ ಒತ್ತಿರಿ.",
   "messagePlay": "ಸಂದೇಶವನ್ನು ಪ್ಲೇ ಮಾಡಿ",
   "messageStop": "ನಿಲ್ಲಿಸಿ",
-  "onboardingSmartRewritingTitle": "ಸ್ಮಾರ್ಟ್ ಮರುಬರವಣಿಗೆ",
   "onboardingSmartRewritingDescription": "ಟೈಲ್‌ ಟ್ಯಾಪ್‌ಗಳನ್ನು ಸಹಜ ವಾಕ್ಯಗಳಾಗಿ ವಿಸ್ತರಿಸಿ ಮತ್ತು ಧಾಟಿಯನ್ನು ಹೊಂದಿಸಿ: ನೇರ, ವೃತ್ತಿಪರ ಅಥವಾ ಸ್ನೇಹಪರ.",
-  "onboardingTranslationTitle": "ಅನುವಾದ",
   "onboardingTranslationDescription": "ಬಹು ಭಾಷೆಗಳಲ್ಲಿ ಸಂವಹನ ನಡೆಸಲು ಬೋರ್ಡ್‌ಗಳನ್ನು ಸ್ಥಳೀಯವಾಗಿ ಅನುವಾದಿಸಿ.",
-  "onboardingPrivacyTitle": "ಖಾಸಗಿ ಮತ್ತು ಆಫ್‌ಲೈನ್",
   "onboardingPrivacyDescription": "ಸಂಪೂರ್ಣವಾಗಿ ನಿಮ್ಮ ಸಾಧನದಲ್ಲಿ ಕಾರ್ಯನಿರ್ವಹಿಸುತ್ತದೆ, ಯಾವುದೇ ಡೇಟಾ ಟ್ರ್ಯಾಕಿಂಗ್ ಇಲ್ಲದೆ ಮತ್ತು ಕ್ಲೌಡ್ ಅವಲಂಬನೆ ಇಲ್ಲದೆ.",
   "onboardingContinue": "ಮುಂದುವರಿಸಿ",
   "errorGenericTitle": "ಏನೋ ತಪ್ಪಾಗಿದೆ",
