+++ messages/ko.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "길게 눌러 메시지를 지웁니다.",
   "messagePlay": "메시지 재생",
   "messageStop": "중지",
-  "onboardingSmartRewritingTitle": "스마트 다시 쓰기",
   "onboardingSmartRewritingDescription": "타일 탭을 자연스러운 문장으로 확장하고 어조를 조정하세요: 직설적인, 전문적인, 친근한.",
-  "onboardingTranslationTitle": "번역",
   "onboardingTranslationDescription": "보드를 기기에서 번역하여 여러 언어로 소통하세요.",
-  "onboardingPrivacyTitle": "비공개 및 오프라인",
   "onboardingPrivacyDescription": "전적으로 기기에서 실행되며, 데이터 추적이 전혀 없고 클라우드에 의존하지 않습니다.",
   "onboardingContinue": "계속",
   "errorGenericTitle": "문제가 발생했습니다",
