+++ messages/nl.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Houd ingedrukt om het bericht te wissen.",
   "messagePlay": "Bericht afspelen",
   "messageStop": "Stoppen",
-  "onboardingSmartRewritingTitle": "Slim herschrijven",
   "onboardingSmartRewritingDescription": "Breid tikken op tegels uit tot natuurlijke zinnen en pas de toon aan: direct, professioneel of vriendelijk.",
-  "onboardingTranslationTitle": "Vertaling",
   "onboardingTranslationDescription": "Vertaal borden lokaal om in meerdere talen te communiceren.",
-  "onboardingPrivacyTitle": "Privé en offline",
   "onboardingPrivacyDescription": "Draait volledig op je apparaat, zonder gegevens te volgen en zonder afhankelijkheid van de cloud.",
   "onboardingContinue": "Doorgaan",
   "errorGenericTitle": "Er is iets misgegaan",
