+++ messages/pl.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Przytrzymaj, aby wyczyścić wiadomość.",
   "messagePlay": "Odtwórz wiadomość",
   "messageStop": "Zatrzymaj",
-  "onboardingSmartRewritingTitle": "Inteligentne przeredagowanie",
   "onboardingSmartRewritingDescription": "Zamień dotknięcia kafelków w naturalne zdania i dostosuj ton: bezpośredni, profesjonalny lub przyjazny.",
-  "onboardingTranslationTitle": "Tłumaczenie",
   "onboardingTranslationDescription": "Tłumacz tablice lokalnie, aby porozumiewać się w wielu językach.",
-  "onboardingPrivacyTitle": "Prywatnie i offline",
   "onboardingPrivacyDescription": "Działa w całości na Twoim urządzeniu, bez śledzenia danych i bez zależności od chmury.",
   "onboardingContinue": "Kontynuuj",
   "errorGenericTitle": "Coś poszło nie tak",
