+++ messages/pt.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Pressione e segure para limpar a mensagem.",
   "messagePlay": "Reproduzir mensagem",
   "messageStop": "Parar",
-  "onboardingSmartRewritingTitle": "Reescrita inteligente",
   "onboardingSmartRewritingDescription": "Transforme os toques nos blocos em frases naturais e ajuste o tom: direto, profissional ou amigável.",
-  "onboardingTranslationTitle": "Tradução",
   "onboardingTranslationDescription": "Traduza as pranchas localmente para se comunicar em vários idiomas.",
-  "onboardingPrivacyTitle": "Privado e offline",
   "onboardingPrivacyDescription": "Funciona inteiramente no seu dispositivo, sem rastreamento de dados nem dependência da nuvem.",
   "onboardingContinue": "Continuar",
   "errorGenericTitle": "Algo deu errado",
