+++ messages/ro.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Apasă lung pentru a șterge mesajul.",
   "messagePlay": "Redă mesajul",
   "messageStop": "Oprește",
-  "onboardingSmartRewritingTitle": "Rescriere inteligentă",
   "onboardingSmartRewritingDescription": "Transformă atingerile pe casete în propoziții naturale și ajustează tonul: direct, profesional sau prietenos.",
-  "onboardingTranslationTitle": "Traducere",
   "onboardingTranslationDescription": "Tradu tablele local pentru a comunica în mai multe limbi.",
-  "onboardingPrivacyTitle": "Privat și offline",
   "onboardingPrivacyDescription": "Rulează în întregime pe dispozitivul tău, fără urmărirea datelor și fără dependență de cloud.",
   "onboardingContinue": "Continuă",
   "errorGenericTitle": "Ceva nu a mers bine",
