+++ messages/ru.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Удерживайте, чтобы очистить сообщение.",
   "messagePlay": "Воспроизвести сообщение",
   "messageStop": "Остановить",
-  "onboardingSmartRewritingTitle": "Умное перефразирование",
   "onboardingSmartRewritingDescription": "Превращайте нажатия на плитки в естественные предложения и настраивайте тон: прямой, профессиональный или дружелюбный.",
-  "onboardingTranslationTitle": "Перевод",
   "onboardingTranslationDescription": "Переводите доски локально, чтобы общаться на нескольких языках.",
-  "onboardingPrivacyTitle": "Приватно и офлайн",
   "onboardingPrivacyDescription": "Работает полностью на вашем устройстве, без отслеживания данных и без зависимости от облака.",
   "onboardingContinue": "Продолжить",
   "errorGenericTitle": "Что-то пошло не так",
