+++ messages/sk.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Dlhým stlačením vymažete správu.",
   "messagePlay": "Prehrať správu",
   "messageStop": "Zastaviť",
-  "onboardingSmartRewritingTitle": "Inteligentný prepis",
   "onboardingSmartRewritingDescription": "Rozšírte ťuknutia na dlaždice na prirodzené vety a upravte tón: priamy, profesionálny alebo priateľský.",
-  "onboardingTranslationTitle": "Preklad",
   "onboardingTranslationDescription": "Prekladajte tabuľky lokálne, aby ste komunikovali vo viacerých jazykoch.",
-  "onboardingPrivacyTitle": "Súkromné a offline",
   "onboardingPrivacyDescription": "Beží úplne vo vašom zariadení, bez sledovania údajov a bez závislosti od cloudu.",
   "onboardingContinue": "Pokračovať",
   "errorGenericTitle": "Niečo sa pokazilo",
