+++ messages/sl.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Za izbris sporočila pridržite.",
   "messagePlay": "Predvajaj sporočilo",
   "messageStop": "Ustavi",
-  "onboardingSmartRewritingTitle": "Pametno preoblikovanje",
   "onboardingSmartRewritingDescription": "Razširite dotike ploščic v naravne stavke in prilagodite ton: neposreden, profesionalen ali prijazen.",
-  "onboardingTranslationTitle": "Prevajanje",
   "onboardingTranslationDescription": "Prevajajte table lokalno za komunikacijo v več jezikih.",
-  "onboardingPrivacyTitle": "Zasebno in brez povezave",
   "onboardingPrivacyDescription": "Deluje v celoti v vaši napravi, brez sledenja podatkom in brez odvisnosti od oblaka.",
   "onboardingContinue": "Nadaljuj",
   "errorGenericTitle": "Nekaj je šlo narobe",
