+++ messages/sv.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Håll inne för att rensa meddelandet.",
   "messagePlay": "Spela upp meddelande",
   "messageStop": "Stoppa",
-  "onboardingSmartRewritingTitle": "Smart omskrivning",
   "onboardingSmartRewritingDescription": "Utöka tryck på rutor till naturliga meningar och justera tonen: direkt, professionell eller vänlig.",
-  "onboardingTranslationTitle": "Översättning",
   "onboardingTranslationDescription": "Översätt tavlor lokalt för att kommunicera på flera språk.",
-  "onboardingPrivacyTitle": "Privat och offline",
   "onboardingPrivacyDescription": "Körs helt på din enhet, utan dataspårning och utan molnberoende.",
   "onboardingContinue": "Fortsätt",
   "errorGenericTitle": "Något gick fel",
