+++ messages/ta.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "செய்தியை அழிக்க நீண்ட நேரம் அழுத்தவும்.",
   "messagePlay": "செய்தியை இயக்கு",
   "messageStop": "நிறுத்து",
-  "onboardingSmartRewritingTitle": "ஸ்மார்ட் மறுஎழுத்து",
   "onboardingSmartRewritingDescription": "டைல் தட்டல்களை இயல்பான வாக்கியங்களாக விரிவாக்கி, தொனியை சரிசெய்யுங்கள்: நேரடி, தொழில்முறை அல்லது நட்பான.",
-  "onboardingTranslationTitle": "மொழிபெயர்ப்பு",
   "onboardingTranslationDescription": "பல மொழிகளில் தொடர்புகொள்ள பலகைகளை சாதனத்திலேயே மொழிபெயர்க்கவும்.",
-  "onboardingPrivacyTitle": "தனிப்பட்டது மற்றும் ஆஃப்லைன்",
   "onboardingPrivacyDescription": "முழுவதுமாக உங்கள் சாதனத்தில் இயங்குகிறது, எந்த தரவுக் கண்காணிப்பும் இல்லாமல், கிளவுட் சார்பும் இல்லாமல்.",
   "onboardingContinue": "தொடரவும்",
   "errorGenericTitle": "ஏதோ தவறு நடந்தது",
