+++ messages/te.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "సందేశాన్ని క్లియర్ చేయడానికి నొక్కి ఉంచండి.",
   "messagePlay": "సందేశాన్ని ప్లే చేయి",
   "messageStop": "ఆపు",
-  "onboardingSmartRewritingTitle": "స్మార్ట్ రీరైటింగ్",
   "onboardingSmartRewritingDescription": "టైల్ ట్యాప్‌లను సహజమైన వాక్యాలుగా విస్తరించి, టోన్‌ను సర్దుబాటు చేయండి: నేరుగా, వృత్తిపరమైన లేదా స్నేహపూర్వక.",
-  "onboardingTranslationTitle": "అనువాదం",
   "onboardingTranslationDescription": "అనేక భాషల్లో సంభాషించేందుకు బోర్డులను స్థానికంగా అనువదించండి.",
-  "onboardingPrivacyTitle": "ప్రైవేట్ & ఆఫ్‌లైన్",
   "onboardingPrivacyDescription": "పూర్తిగా మీ పరికరంలోనే పనిచేస్తుంది, ఎలాంటి డేటా ట్రాకింగ్ లేకుండా, క్లౌడ్ ఆధారపడటం లేకుండా.",
   "onboardingContinue": "కొనసాగించు",
   "errorGenericTitle": "ఏదో తప్పు జరిగింది",
