+++ messages/th.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "กดค้างเพื่อล้างข้อความ",
   "messagePlay": "เล่นข้อความ",
   "messageStop": "หยุด",
-  "onboardingSmartRewritingTitle": "การเขียนใหม่อัจฉริยะ",
   "onboardingSmartRewritingDescription": "ขยายการแตะไทล์ให้เป็นประโยคที่เป็นธรรมชาติ และปรับโทน: ตรงไปตรงมา เป็นทางการ หรือเป็นมิตร",
-  "onboardingTranslationTitle": "การแปล",
   "onboardingTranslationDescription": "แปลบอร์ดในเครื่อง เพื่อสื่อสารได้หลายภาษา",
-  "onboardingPrivacyTitle": "ส่วนตัวและออฟไลน์",
   "onboardingPrivacyDescription": "ทำงานบนอุปกรณ์ของคุณทั้งหมด ไม่มีการติดตามข้อมูลและไม่ต้องพึ่งพาคลาวด์",
   "onboardingContinue": "ดำเนินการต่อ",
   "errorGenericTitle": "เกิดข้อผิดพลาดบางอย่าง",
