+++ messages/tr.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Mesajı temizlemek için uzun basın.",
   "messagePlay": "Mesajı oynat",
   "messageStop": "Durdur",
-  "onboardingSmartRewritingTitle": "Akıllı yeniden yazma",
   "onboardingSmartRewritingDescription": "Karelere dokunuşları doğal cümlelere dönüştürün ve tonu ayarlayın: doğrudan, profesyonel veya samimi.",
-  "onboardingTranslationTitle": "Çeviri",
   "onboardingTranslationDescription": "Birden fazla dilde iletişim kurmak için panoları yerel olarak çevirin.",
-  "onboardingPrivacyTitle": "Gizli ve çevrimdışı",
   "onboardingPrivacyDescription": "Tamamen cihazınızda çalışır; veri takibi olmadan ve buluta bağımlılık olmadan.",
   "onboardingContinue": "Devam et",
   "errorGenericTitle": "Bir şeyler ters gitti",
