+++ messages/uk.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Утримуйте, щоб очистити повідомлення.",
   "messagePlay": "Відтворити повідомлення",
   "messageStop": "Зупинити",
-  "onboardingSmartRewritingTitle": "Розумне переписування",
   "onboardingSmartRewritingDescription": "Перетворюйте натискання на плитки на природні речення та налаштовуйте тон: прямий, професійний або дружній.",
-  "onboardingTranslationTitle": "Переклад",
   "onboardingTranslationDescription": "Перекладайте дошки локально, щоб спілкуватися кількома мовами.",
-  "onboardingPrivacyTitle": "Приватно й офлайн",
   "onboardingPrivacyDescription": "Працює повністю на вашому пристрої, без відстеження даних і без залежності від хмари.",
   "onboardingContinue": "Продовжити",
   "errorGenericTitle": "Щось пішло не так",
