+++ messages/vi.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "Nhấn giữ để xóa tin nhắn.",
   "messagePlay": "Phát tin nhắn",
   "messageStop": "Dừng",
-  "onboardingSmartRewritingTitle": "Viết lại thông minh",
   "onboardingSmartRewritingDescription": "Mở rộng các lần chạm ô thành câu tự nhiên và điều chỉnh giọng điệu: trực tiếp, chuyên nghiệp hoặc thân thiện.",
-  "onboardingTranslationTitle": "Dịch thuật",
   "onboardingTranslationDescription": "Dịch bảng ngay trên thiết bị để giao tiếp bằng nhiều ngôn ngữ.",
-  "onboardingPrivacyTitle": "Riêng tư & ngoại tuyến",
   "onboardingPrivacyDescription": "Chạy hoàn toàn trên thiết bị của bạn, không theo dõi dữ liệu và không phụ thuộc vào đám mây.",
   "onboardingContinue": "Tiếp tục",
   "errorGenericTitle": "Đã xảy ra lỗi",
