+++ messages/zh.json
@@ -139,11 +139,8 @@
   "messageBackspaceLongPressHint": "长按可清空消息。",
   "messagePlay": "播放消息",
   "messageStop": "停止",
-  "onboardingSmartRewritingTitle": "智能改写",
   "onboardingSmartRewritingDescription": "将点按图块扩展为自然的句子，并调整语气：直接、专业或友好。",
-  "onboardingTranslationTitle": "翻译",
   "onboardingTranslationDescription": "在本地翻译沟通板，以使用多种语言交流。",
-  "onboardingPrivacyTitle": "私密且离线",
   "onboardingPrivacyDescription": "完全在你的设备上运行，零数据追踪，也不依赖云端。",
   "onboardingContinue": "继续",
   "errorGenericTitle": "出了点问题",
