+++ src/shared/components/external-link.tsx
@@ -2,10 +2,11 @@ import Link, { type LinkProps } from "@mui/material/Link";
 import visuallyHidden from "@mui/utils/visuallyHidden";
 import { m } from "@paraglide/messages.js";
 import { useTranslate } from "@shared/language/use-translate";
+import { mergeSx } from "@shared/theme/merge-sx";
 
 type ExternalLinkProps = Omit<LinkProps, "target" | "rel">;
 
-export function ExternalLink({ children, ...props }: ExternalLinkProps) {
+export function ExternalLink({ children, sx, ...props }: ExternalLinkProps) {
   const t = useTranslate();
 
   return (
@@ -14,7 +15,7 @@ export function ExternalLink({ children, ...props }: ExternalLinkProps) {
       target="_blank"
       rel="noopener noreferrer"
       underline="always"
-      sx={{ whiteSpace: "nowrap" }}
+      sx={mergeSx({ whiteSpace: "nowrap" }, sx)}
     >
       {children}
       <span style={visuallyHidden}> {t(m.opensInNewTab)}</span>
